refactor: improve the code
