+++ pages/_app.tsx
@@ -1,6 +1,7 @@
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
-import Layout from '@/components/Layout'
+
+import { Layout } from '@/components'
 
 import '@styles/globals.scss'
 
@@ -8,8 +9,8 @@ export default function App({ Component, pageProps }: AppProps) {
   return (
     <Layout>
       <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
+        <title>View Profile | Github</title>
+        <meta name="description" content="View your profile of github" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
